# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Nanaya",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token from /users/login"
      }
    }
  },
  "paths": {
    "/comments/{documentId}": {
      "get": {
        "summary": "Get all comments for a document",
        "description": "Retrieve all comments associated with a specific document, sorted by creation date (newest first)",
        "tags": [
          "Comments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "documentId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the document"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of comments for the document",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "string",
                        "description": "Comment ID"
                      },
                      "documentId": {
                        "type": "string",
                        "description": "Document ID"
                      },
                      "userId": {
                        "type": "object",
                        "description": "The comment author",
                        "properties": {
                          "_id": {
                            "type": "string"
                          },
                          "screen_name": {
                            "type": "string"
                          },
                          "teacher": {
                            "type": "boolean"
                          }
                        }
                      },
                      "text": {
                        "type": "string",
                        "description": "Comment text content"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Comment creation timestamp"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - no valid token provided"
          },
          "500": {
            "description": "Failed to fetch comments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Failed to fetch comments"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/comments": {
      "post": {
        "summary": "Add a comment to a document",
        "description": "Create a new comment on a specific document",
        "tags": [
          "Comments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "documentId",
                  "text"
                ],
                "properties": {
                  "documentId": {
                    "type": "string",
                    "description": "ID of the document to comment on",
                    "example": "507f1f77bcf86cd799439011"
                  },
                  "text": {
                    "type": "string",
                    "description": "The comment text",
                    "example": "This document needs revision"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Comment created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string",
                      "description": "Comment ID"
                    },
                    "documentId": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "string"
                    },
                    "text": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing documentId or text",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Missing documentId or text"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - no valid token provided"
          },
          "500": {
            "description": "Failed to create comment",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Failed to create comment"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/comments/{commentId}": {
      "delete": {
        "summary": "Delete a comment",
        "description": "Delete a comment from a document. Only the document owner can delete comments, but cannot delete comments written by teachers.",
        "tags": [
          "Comments"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "commentId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the comment to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Comment deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Comment deleted successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - not the document owner",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Unauthorized"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - cannot delete comments written by teachers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Cannot delete comments written by teachers"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Comment or document not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Comment not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Failed to delete comment",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Failed to delete comment"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}